+++ python/tests/unittests/test_vqe.py
@@ -146,10 +146,10 @@ def test_vqe_two_qubit_list(optimizer, kernel_two_qubit_vqe_list,
     """
     # Should be able to call this by passing a function that returns a kernel:
     got_expectation, got_parameters = cudaq.vqe(kernel_two_qubit_vqe_list,
-                                               hamiltonian_2q,
-                                               optimizer,
-                                               parameter_count=1,
-                                               shots=-1)
+                                                hamiltonian_2q,
+                                                optimizer,
+                                                parameter_count=1,
+                                                shots=-1)
 
     # Known minimal expectation value for this system:
     want_expectation_value = -1.7487948611472093
@@ -232,12 +232,13 @@ def test_vqe_two_qubit_list_gradients(optimizer, gradient,
     `Kernel` and `Callable` overloads.
     """
     # Should be able to call this by passing a function that returns a kernel:
-    got_expectation, got_parameters = cudaq.vqe(kernel=kernel_two_qubit_vqe_list,
-                                               gradient_strategy=gradient,
-                                               spin_operator=hamiltonian_2q,
-                                               optimizer=optimizer,
-                                               parameter_count=1,
-                                               shots=-1)
+    got_expectation, got_parameters = cudaq.vqe(
+        kernel=kernel_two_qubit_vqe_list,
+        gradient_strategy=gradient,
+        spin_operator=hamiltonian_2q,
+        optimizer=optimizer,
+        parameter_count=1,
+        shots=-1)
 
     # Known minimal expectation value for this system:
     want_expectation_value = -1.7487948611472093
